moves checks to manual isuserauth checks
